Alterações - Diretório API e Novas páginas
