+++ errorcenter/errorcenter/urls.py
@@ -18,5 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('errors.urls'))
+    path('api/', include('errors.api.urls')),
+    path('', include('errors.urls')),
 ]
