+++ errorcenter/errors/api/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import User, Error
+from ..models import User, Error
 
 
 class UserSerializer(serializers.ModelSerializer):
