+++ errorcenter/errors/templates/errors/error_list.html
@@ -1,0 +1,6 @@
+
+{% for error in errors %}
+<h2>{{ error.title }}</h2>
+{% empty %}
+<h1>Não há produtos para serem listados.</h1>
+{% endfor %}
