+++ errorcenter/errors/templates/errors/user_register.html
@@ -1,0 +1,1 @@
+<h1>Este é um formulário de cadastro de usuários.</h1>
